Start and finised
Made it so when hovered over the stars it changes the number of stars highlighted and the count of how many highlighted. Then when it is clicked it locks it. Also counts up the number of characters in the text box and applies that number to the character box on the HTML page.